Move setup_logger into main
This may not be the best place but means the private CLI can avoid importing utils.py
