+++ pyani_plus/__init__.py
@@ -28,6 +28,9 @@ methods. It is a reimplemented version of ``pyani`` with support for
 additional schedulers and methods.
 """
 
+import logging
+from pathlib import Path
+
 from rich.progress import (
     BarColumn,
     MofNCompleteColumn,
@@ -50,3 +53,33 @@ PROGRESS_BAR_COLUMNS = [
     # Add this last as have some out of N and some out of N^2:
     MofNCompleteColumn(),
 ]
+
+
+def setup_logger(log_folder: Path | None, name: str) -> logging.Logger:
+    """Return a file-based logger alongside a Rich console logger.
+
+    The file logger defaults to DEBUG level, but the less verbose INFO for the terminal.
+    """
+    logger = logging.getLogger(f"{__package__}.{name}")
+    logger.setLevel(logging.DEBUG)
+    if logger.hasHandlers():  # remove all previous handlers to avoid duplicate entries
+        logger.handlers.clear()
+
+    if log_folder == Path("-"):
+        logger.debug("Not logging to file.")
+        return logger
+
+    filename = (log_folder / (name + ".log")) if log_folder else Path(name + ".log")
+    file_handler = logging.FileHandler(filename, mode="a")
+    file_handler.setLevel(logging.DEBUG)
+
+    fmt = "%(asctime)s %(levelname)9s %(filename)21s:%(lineno)-3s | %(message)s"
+    formatter = logging.Formatter(fmt=fmt, datefmt="%Y-%m-%d %H:%M:%S")
+    file_handler.setFormatter(formatter)
+
+    logger.addHandler(file_handler)
+
+    msg = f"Logging to {filename}"
+    logger.info(msg)  # Want this to appear on the terminal
+
+    return logger
