+++ pyani_plus/utils.py
@@ -23,7 +23,6 @@
 
 import gzip
 import hashlib
-import logging
 import os
 import shutil
 import subprocess
@@ -314,33 +313,3 @@ def stage_file(
     else:
         # Make a symlink pointing to the original
         staged_filename.symlink_to(input_filename)
-
-
-def setup_logger(log_folder: Path | None, name: str) -> logging.Logger:
-    """Return a file-based logger alongside a Rich console logger.
-
-    The file logger defaults to DEBUG level, but the less verbose INFO for the terminal.
-    """
-    logger = logging.getLogger(f"{__package__}.{name}")
-    logger.setLevel(logging.DEBUG)
-    if logger.hasHandlers():  # remove all previous handlers to avoid duplicate entries
-        logger.handlers.clear()
-
-    if log_folder == Path("-"):
-        logger.debug("Not logging to file.")
-        return logger
-
-    filename = (log_folder / (name + ".log")) if log_folder else Path(name + ".log")
-    file_handler = logging.FileHandler(filename, mode="a")
-    file_handler.setLevel(logging.DEBUG)
-
-    fmt = "%(asctime)s %(levelname)9s %(filename)21s:%(lineno)-3s | %(message)s"
-    formatter = logging.Formatter(fmt=fmt, datefmt="%Y-%m-%d %H:%M:%S")
-    file_handler.setFormatter(formatter)
-
-    logger.addHandler(file_handler)
-
-    msg = f"Logging to {filename}"
-    logger.info(msg)  # Want this to appear on the terminal
-
-    return logger
